% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_ccphate.R
\name{ccphate}
\alias{ccphate}
\title{Consensus Clustering with PHATE Geometry}
\usage{
ccphate(data, k = 2, ...)
}
\arguments{
\item{data}{an \eqn{(n\times p)} matrix of row-stacked observations.}

\item{k}{the number of clusters (default: 2).}

\item{...}{extra parameters including \describe{
\item{nnbd}{size of nearest neighborhood (default: 5).}
\item{alpha}{decay parameter for Gaussian kernel exponent (default: 2).}
\item{ratio}{percentage of samples to be used for each resampling run with the minimum is set as 0.25 (default: 0.9).}
\item{niter}{number of resampling runs where the minimum is set 50 (default: 100).}
}}
}
\description{
Consensus Clustering with PHATE Geometry
}
\concept{algorithm}
