% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_consensus_auc.R
\name{consensus.auc}
\alias{consensus.auc}
\title{Consensus Determination of K : Original Method}
\usage{
consensus.auc(clist)
}
\arguments{
\item{clist}{a list of consensus matrices wrapped as \code{T4cluster:consensus} objects.}
}
\value{
a \code{data.frame} consisting of \describe{
\item{\code{nclust}}{numbers of clusters.}
\item{\code{delta}}{delta values as noted in the paper.}
}
}
\description{
Consensus Determination of K : Original Method
}
\examples{
\donttest{

}



}
\concept{soc}
