% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_consensus_pac.R
\name{consensus.pac}
\alias{consensus.pac}
\title{Consensus Determination of K : PAC method}
\usage{
consensus.pac(clist, lb = 0.1, ub = 0.9)
}
\description{
Consensus Determination of K : PAC method
}
\concept{soc}
