% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace_EKSS.R
\name{EKSS}
\alias{EKSS}
\title{Ensemble of K-Subspaces}
\usage{
EKSS(data, k = 2, d = 2, q = floor(nrow(data) * 0.75), B = 500, iter = 0)
}
\description{
Ensemble of K-Subspaces
}
\examples{
\donttest{
## generate a toy example
set.seed(10)
tester = genLP(n=100, nl=2, np=1, iso.var=0.1)
data   = tester$data
label  = tester$class

## do PCA for data reduction
proj = base::eigen(stats::cov(data))$vectors[,1:2]
dat2 = data\%*\%proj

## run EKSS algorithm with k=2,3,4 with EKSS-0 and 5 iteration
out2zero = EKSS(data, k=2)
out3zero = EKSS(data, k=3)
out4zero = EKSS(data, k=4)

out2iter = EKSS(data, k=2, iter=10)
out3iter = EKSS(data, k=3, iter=10)
out4iter = EKSS(data, k=4, iter=10)

## extract label information
lab2zero = out2zero$cluster
lab3zero = out3zero$cluster
lab4zero = out4zero$cluster

lab2iter = out2iter$cluster
lab3iter = out3iter$cluster
lab4iter = out4iter$cluster

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(2,3))
plot(dat2, pch=19, cex=0.9, col=lab2zero, main="LRSC Relaxed:K=2")
plot(dat2, pch=19, cex=0.9, col=lab3zero, main="LRSC Relaxed:K=3")
plot(dat2, pch=19, cex=0.9, col=lab4zero, main="LRSC Relaxed:K=4")
plot(dat2, pch=19, cex=0.9, col=lab2iter, main="LRSC Exact:K=2")
plot(dat2, pch=19, cex=0.9, col=lab3iter, main="LRSC Exact:K=3")
plot(dat2, pch=19, cex=0.9, col=lab4iter, main="LRSC Exact:K=4")
par(opar)
}

}
\concept{subspace}
