% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_gbphate.R
\name{gbphate}
\alias{gbphate}
\title{Generalized Bayesian Clustering with PHATE Geometry}
\usage{
gbphate(data, k = 2, ...)
}
\arguments{
\item{data}{an \eqn{(n\times p)} matrix of row-stacked observations.}

\item{k}{the number of clusters (default: 2).}

\item{...}{extra parameters including \describe{
\item{nnbd}{size of nearest neighborhood (default: 5).}
\item{alpha}{decay parameter for Gaussian kernel exponent (default: 10).}
}}
}
\description{
Generalized Bayesian Clustering with PHATE Geometry
}
\concept{algorithm}
