% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_gmm11R.R
\name{gmm11R}
\alias{gmm11R}
\title{Regularized GMM by Ruan et al. (2011)}
\usage{
gmm11R(data, k = 2, lambda = 1, ...)
}
\arguments{
\item{data}{an \eqn{(n\times p)} matrix of row-stacked observations.}

\item{k}{the number of clusters (default: 2).}

\item{lambda}{regularization parameter for graphical lasso (default: 1).}

\item{...}{extra parameters including \describe{
\item{maxiter}{the maximum number of iterations (default: 10).}
\item{nstart}{the number of random initializations (default: 5).}
\item{usediag}{a logical; covariances are diagonal if \code{TRUE}, or full covariances are returned for \code{FALSE} (default: \code{FALSE}).}
}}
}
\value{
a named list of S3 class \code{T4cluster} containing 
\describe{
\item{cluster}{a length-\eqn{n} vector of class labels (from \eqn{1:k}).}
\item{mean}{a \eqn{(k\times p)} matrix where each row is a class mean.}
\item{variance}{a \eqn{(p\times p\times k)} array where each slice is a class covariance.}
\item{weight}{a length-\eqn{k} vector of class weights that sum to 1.}
\item{loglkd}{log-likelihood of the data for the fitted model.}
\item{algorithm}{name of the algorithm.}
}
}
\description{
Ruan et al. (2011) proposed a regularized covariance estimation by 
graphical lasso to cope with high-dimensional scenario where conventional 
GMM might incur singular covariance components. Authors proposed to use 
\eqn{\lambda} as a regularization parameter as normally used in 
sparse covariance/precision estimation problems and suggested to use the 
model with the smallest BIC values. Consult \code{\link{gmmscore}} for 
information criteria for model selection.
}
\examples{
# -------------------------------------------------------------
#            clustering with 'iris' dataset
# -------------------------------------------------------------
## PREPARE
data(iris)
X   = as.matrix(iris[,1:4])
lab = as.integer(as.factor(iris[,5]))

## EMBEDDING WITH PCA
X2d = Rdimtools::do.pca(X, ndim=2)$Y  

## COMPARE WITH STANDARD GMM
cl.gmm = gmm(X, k=3)$cluster
cl.11Rf = gmm11R(X, k=3)$cluster
cl.11Rd = gmm11R(X, k=3, usediag=TRUE)$cluster

## VISUALIZATION
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3), pty="s")
plot(X2d, col=cl.gmm,  pch=19, main="standard GMM")
plot(X2d, col=cl.11Rf, pch=19, main="gmm11R: full covs")
plot(X2d, col=cl.11Rd, pch=19, main="gmm11R: diagonal covs")
par(opar)

}
\references{
\insertRef{ruan_regularized_2011}{T4cluster}
}
\concept{algorithm}
