% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_gmmscore.R
\name{gmmscore}
\alias{gmmscore}
\title{Compute Information Criterion for Fitted GMM Models}
\usage{
gmmscore(gmmobj)
}
\arguments{
\item{gmmobj}{an output of any GMM routines in our package of \code{T4cluster} class.}
}
\value{
a vector of information criteria scores.
}
\description{
Model selection in GMM can be done in many ways, including Information Criterion (IC). 
Currently, we currently support Akaike (AIC), Bayesian (BIC), and Hannan-Quinn (HQC) scores 
for comparing models where a smaller value indicates a better model.
}
\examples{
# -------------------------------------------------------------
#            clustering with 'iris' dataset
# -------------------------------------------------------------
## PREPARE
data(iris)
X   = as.matrix(iris[,1:4])

## FIT THE MODELS
models = list()
for (i in 1:5){
  models[[i]] = gmm(X, k=i)
}

## COMPUTE SCORES
scores = c()
for (i in 1:5){
  scores = rbind(scores, gmmscore(models[[i]]))
}

## VISUALIZE
opar <- par(no.readonly=TRUE)
matplot(scores, main="Information Criteria", type="b", 
        xlab="number of clusters", pch=letters[1:3])
legend("bottomleft", c("(a) AIC","(b) BIC","(c) HQC"), col=1:3, fill=1:3)
par(opar)

}
\references{
\insertRef{akaike_new_1974}{T4cluster}

\insertRef{schwarz_estimating_1978}{T4cluster}

\insertRef{hannan_determination_1979}{T4cluster}
}
\concept{utility}
