% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace_msm.R
\name{predict.msm}
\alias{predict.msm}
\title{S3 method to predict class label of new data with 'msm' object}
\usage{
\method{predict}{msm}(object, newdata, ...)
}
\arguments{
\item{object}{an \code{'msm'} object from \code{\link{msm}} function.}

\item{newdata}{an \eqn{(m\times p)} matrix of row-stacked observations.}

\item{...}{extra parameters (not necessary).}
}
\value{
a length-\eqn{m} vector of class labels.
}
\description{
Given an instance of \code{msm} class from \code{\link{msm}} function, predict 
class label of a new data.
}
\concept{utility}
