% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace_MSM.R
\name{predict.MSM}
\alias{predict.MSM}
\title{S3 method to predict class label of new data with 'MSM' object}
\usage{
\method{predict}{MSM}(object, newdata, ...)
}
\arguments{
\item{object}{an \code{'MSM'} object from \code{\link{MSM}} function.}

\item{newdata}{an \eqn{(m\times p)} matrix of row-stacked observations.}

\item{...}{extra parameters (not necessary).}
}
\value{
a length-\eqn{m} vector of class labels.
}
\description{
Given an instance of \code{MSM} class from \code{\link{MSM}} function, predict 
class label of a new data.
}
\concept{subspace}
